requirements/vector_db_requirements.txt
# 向量資料庫整合模組依賴
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.18
weaviate-client>=3.24.0
pymilvus>=2.3.0
qdrant-client>=1.6.0
numpy>=1.24.0
tqdm>=4.66.0
pydantic>=2.0.0
python-dotenv>=1.0.0
